Remove config::float_type
- Delete tests/experimental
- Remove config::float_type and change all usages of float_type
  and replace with "float" type.
- Remove test_package tweaks file as the defaults will suffice
- Add conanfiles to gitignore
- Combine double & float tests back into a single unit test

Resolves #405

+++ include/libhal/percentage.hpp
@@ -71,7 +71,7 @@ public:
   }
   constexpr bool equals(percentage_t p_other, float_t p_epsilon = 1e-9f) const
   {
-    return hal::equals(m_value, p_other.value(), p_epsilon);
+    return hal::equals(m_value, p_other.value(), static_cast<float>(p_epsilon));
   }
   constexpr bool operator<=>(const percentage_t& p_percent) const = default;
   constexpr bool operator==(const percentage_t& p_percent) const
